pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.3.0",
    "flask-mail>=0.10.0",
    "flask>=3.1.2",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "flask-migrate>=4.1.0",
    "wtforms>=3.2.1",
    "flask-login>=0.6.3",
    "flask-wtf>=1.2.2",
    "pillow>=11.3.0",
    "bleach>=6.2.0",
    "sqlalchemy>=2.0.43",
    "werkzeug>=3.1.3",
    "pyjwt>=2.10.1",
]
